+++ lib/active_admin/graphql.rb
@@ -77,6 +77,7 @@ module ActiveAdmin
           graphql/run_action_mutation_dsl
           graphql/key_value_pair_input
           graphql/policy_set_cache
+          graphql/mutation_execution_error
           graphql/schema_builder
         ].each { |path| require_relative path }
       end